change aws-java-sdk to be provided
Currently this module depends on the aws-java-sdk root
project that brings in 70+ sub-jars for various services.
In most cases a particular user will only need a small
subset of these jars.

With this change the aws-java-sdk is now provided so the
user would be expected to explicitly depend on the jars
for the services they need. In general that should already
be the case because the user should have a direct dependency
if they are using the model objects needed to work with
these mappers.

One downside to this change is the user will likely need
to be more diligent about ensuring that the awsobjectmapper
version matches the aws-java-sdk version.