update to 1.11.469
The size of the configure method grew to large causing
compilation errors:

```
generated/com/netflix/awsobjectmapper/AmazonObjectMapperConfigurer.java:32: error: code too large
public static void configure(ObjectMapper objectMapper) {
                   ^
```

For now we just put in a pattern to ignore some of the
services we do not use that generate a lot of mixins.
Ideally it will get refactored later to generate the
method differently so we avoid the limit.