aws-java-sdk 1.11.556 (#60)
Update to aws-java-sdk 1.11.556. This update caused us to
hit a `code too large` error again. See #51 for context on the
last time this happened. The dependencies have been pruned
to only include the SDK projects that are known to be used
with Edda or related projects.

Also switches the Travis build to openjdk8, it was failing to
find oraclejdk8.